arm64: dts: cn9130-malibu-gw8901: add gpio controlled FAN
Enable a GPIO controlled FAN which is suited well to support PWM fans.

This is configured for a FAN with cooling levels at 0%, 40%, 66%, and 100%
with two TACH pulses per revoluation.

This is mapped to the ap_thermal_ic thermal zone which already has a 100C
critical trip.

Configure for active cooling to be on 100% before passive cooling (at
85C from the ap_thermal_cpu* cooling zones).

Signed-off-by: Tim Harvey <tharvey@gateworks.com>